+++ NEWS.md
@@ -1,4 +1,12 @@
 
+# tern 0.5.1
+
+## Package Internal Changes
+
+* added function `compound_table`
+* In forest plot functions, added formatting to display extreme values to ">999.9"." 
+* added `t_ae_ctc` adverse events table by highest NCI CTCAE grade function.
+
 # tern 0.5.0
 
 First version where analysis functions names and arguments have been harmonized.